+++ src/main/java/com/team/delightserver/exception/web/handler/GlobalExceptionHandler.java
@@ -1,0 +1,15 @@
+package com.team.delightserver.exception.web.handler;
+
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @Created by Doe
+ * @Date: 2021/08/30
+ */
+
+@ControllerAdvice
+@RestController
+public class GlobalExceptionHandler {
+
+}
